requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-genai>=0.2.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
numpy>=1.26.0
pyaudio>=0.2.14
google-cloud-texttospeech>=2.14.0
